ai/foodApi/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
pillow